pyproject.toml
[tool.poetry]
name = "vocabinterval"
version = "0.1.0"
description = ""
authors = ["Chanov Sergey <s.c.chanov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.3.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
